feat: add reusable drop zone
